requirements.txt
autocorrect==2.6.1
certifi==2021.10.8
chardet==3.0.4
click==8.1.2
colorama==0.4.4
Flask==2.1.1
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.1
lxml==4.8.0
MarkupSafe==2.1.1
python-docx==0.8.11
python-dotenv==0.20.0
rfc3986==1.5.0
sniffio==1.2.0
Werkzeug==2.1.1
